ServiceView - FormView that handles mechanics of calling a Service
Follows the usual CBV pattern of most data coming from method calls that
can be overridden as necessary.  `form_valid` mimics the FBV example in
the README.

Possibly controversial is `get_form_class` that will use the Service as
the Form if no form_class is given.  My homegrown similar style uses a
Django Form as both the UI and business rules (via a execute() instance
method) which is why I coded ServiceView that way.  But for this
library, I could go either way.